% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runchecks.R
\docType{data}
\name{all_packages}
\alias{all_packages}
\title{Quality information for phyogenetics packages}
\format{A list containing lists of quality information}
\usage{
all_packages
}
\description{
A dataset of outputs of run_package_check for every package in the Phylogenetics task view
}
\keyword{datasets}
