% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runchecks.R
\name{cache_all_from_view}
\alias{cache_all_from_view}
\title{Runs tests for all packages in a task view and stores inside this package}
\usage{
cache_all_from_view(view = "Phylogenetics",
  ncores = parallel::detectCores())
}
\arguments{
\item{view}{View name to get packages from}

\item{ncores}{Number of cores to use}
}
\value{
nothing, though a data file has been saved
}
\description{
Runs tests for all packages in a task view and stores inside this package
}
