% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runchecks.R
\name{check_all_packages}
\alias{check_all_packages}
\title{Check a vector of packages}
\usage{
check_all_packages(pkgs, ncores = 1)
}
\arguments{
\item{pkgs}{a vector of packages}

\item{ncores}{how many cores to use}
}
\value{
a list of of package checks, one per package
}
\description{
This will run the checks for all the packages in the vector. You can run in parallel if you set ncores to greater than one and have an operating system that can support this. When this is done, you can use sapply(<list of packages>, summarize_goodness) to get summaries for all.
}
\examples{
all_packages <- get_all_packages_in_view("Phylogenetics")
fewer_packages <- all_packages[1:3]
all_checks <- check_all_packages(fewer_packages)
organized_results <- t(sapply(all_checks, summarize_goodness))
print(organized_results)
}
