% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runchecks.R
\name{get_all_packages_in_view}
\alias{get_all_packages_in_view}
\title{Get names of all packages from a view}
\usage{
get_all_packages_in_view(view)
}
\arguments{
\item{view}{View name to get packages from}
}
\value{
vector of packages
}
\description{
Uses very elegant code by Jonathan Chang: https://jonathanchang.org/blog/what-r-package-for-phylogenetics-is-the-most-popular/
}
