% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runchecks.R
\name{plot_package_check}
\alias{plot_package_check}
\title{Plot downloads and package updates}
\usage{
plot_package_check(pkgcheck)
}
\arguments{
\item{pkgcheck}{output from run_package_check}
}
\value{
A ggplot2 object
}
\description{
This will plot downloads, a smoothed line of downloads through time, and vertical lines showing package updates on CRAN
}
