% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runchecks.R
\name{run_package_check}
\alias{run_package_check}
\title{Run package check}
\usage{
run_package_check(pkg)
}
\arguments{
\item{pkg}{The package name (a string)}
}
\value{
a list containing a goodpractice object (goodpractice_result), a data.frame from packagemetrics packagemetrics_result), a data.frame of other information from crandb (package_info), and output from cran_donloads of daily downloads since 2013
}
\description{
This runs a check of a given package by downloading it, running goodpractice, and checking other aspects of it.
}
\examples{
goodness <- run_package_check('hisse')
print(goodness$goodpractice_result)
}
