% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runchecks.R
\name{summarize_goodness}
\alias{summarize_goodness}
\title{Summarize goodness of a package}
\usage{
summarize_goodness(pkgcheck)
}
\arguments{
\item{pkgcheck}{output from run_package_check}
}
\value{
A data.frame with information
}
\description{
Gives useful information about downloads, updates, best practices, and more
}
\examples{
 t(sapply(all_packages, summarize_goodness))
}
